split NodeRuntime, make single triggerProposeF, remove rspaceSheduler
