+++ node/src/main/scala/coop/rchain/node/instances/BlockProcessorInstance.scala
@@ -1,32 +1,24 @@
 package coop.rchain.node.instances
 
 import cats.effect.Concurrent
-import cats.effect.concurrent.{Deferred, Ref}
+import cats.effect.concurrent.Ref
 import cats.instances.list._
 import cats.syntax.all._
 import coop.rchain.casper.blocks.BlockProcessor
-import coop.rchain.casper.blocks.proposer.{BugError, ProposeResult}
 import coop.rchain.casper.protocol.BlockMessage
-import coop.rchain.casper.{
-  BlockError,
-  BlockStatus,
-  Casper,
-  PrettyPrinter,
-  ValidBlock,
-  ValidBlockProcessing
-}
+import coop.rchain.casper.{Casper, PrettyPrinter, ValidBlockProcessing}
 import coop.rchain.models.BlockHash.BlockHash
+import coop.rchain.node.NodeRuntime.ProposeFunction
 import coop.rchain.shared.Log
-import fs2.concurrent.Queue
 import fs2.Stream
+import fs2.concurrent.Queue
 
 object BlockProcessorInstance {
   def create[F[_]: Concurrent: Log](
       blocksQueue: Queue[F, (Casper[F], BlockMessage)],
       blockProcessor: BlockProcessor[F],
       state: Ref[F, Set[BlockHash]],
-      proposerQueue: Queue[F, (Casper[F], Deferred[F, Option[Int]])],
-      autopropose: Boolean
+      triggerProposeF: Option[ProposeFunction[F]]
   ): Stream[F, (Casper[F], BlockMessage, ValidBlockProcessing)] = {
 
     // Node can handle `parallelism` blocks in parallel, or they will be queued
@@ -90,11 +82,7 @@ object BlockProcessorInstance {
               } yield ()
             }
             .evalTap { v =>
-              // we don't care about result of propose here, whether its started or not.
-              // so let deferred be not used
-              (Deferred[F, Option[Int]] >>= { d =>
-                proposerQueue.enqueue1((v._1, d))
-              }).whenA(autopropose)
+              triggerProposeF.traverse(f => f(v._1))
             }
             // ensure to remove hash from state
             .onFinalize(state.update(s => s - b.blockHash))
